+++ Runtime/ArenaClient.cs
@@ -31,7 +31,7 @@ namespace ArenaUnity
     /// Class to manage a singleton instance of the ARENA client connection.
     /// </summary>
     [HelpURL("https://arena.conix.io")]
-    [DisallowMultipleComponent()]
+    [DisallowMultipleComponent]
     [AddComponentMenu("ArenaClient", 0)]
     public class ArenaClient : M2MqttUnityClient
     {
@@ -141,6 +141,8 @@ namespace ArenaUnity
         protected override void Start()
         {
             StartCoroutine(SceneSignin());
+
+            Selection.activeGameObject = gameObject; // client focus at runtime starts
         }
 
         // Update is called once per frame
@@ -281,24 +283,24 @@ namespace ArenaUnity
             ArenaClientTransform = FindObjectOfType<ArenaClient>().transform;
             string jsonString = cd.result.ToString();
             JArray jsonVal = JArray.Parse(jsonString);
-            dynamic objects = jsonVal;
+            dynamic persistMessages = jsonVal;
             // establish objects
             int objects_num = 1;
             if (Directory.Exists(Application.dataPath + "/ArenaUnity"))
                 Directory.Delete(Application.dataPath + "/ArenaUnity", true);
             if (File.Exists(Application.dataPath + "/ArenaUnity.meta"))
                 File.Delete(Application.dataPath + "/ArenaUnity.meta");
-            foreach (dynamic obj in objects)
+            foreach (dynamic msg in persistMessages)
             {
-                DisplayCancelableProgressBar("ARENA Persistance", $"Loading object-id: {(string)obj.object_id}", objects_num / (float)jsonVal.Count);
+                DisplayCancelableProgressBar("ARENA Persistance", $"Loading object-id: {(string)msg.object_id}", objects_num / (float)jsonVal.Count);
                 string localPath = null;
-                if (isElement(obj.attributes) && isElement(obj.attributes.url) && !isElementEmpty(obj.attributes.url))
+                if (isElement(msg.attributes) && isElement(msg.attributes.url) && !isElementEmpty(msg.attributes.url))
                 {
-                    cd = new CoroutineWithData(this, DownloadAssets(obj));
+                    cd = new CoroutineWithData(this, DownloadAssets((string)msg.type, msg.attributes));
                     yield return cd.coroutine;
                     localPath = cd.result.ToString();
                 }
-                CreateUpdateObject((string)obj.object_id, (string)obj.type, obj.attributes, localPath);
+                CreateUpdateObject((string)msg.object_id, (string)msg.type, msg.attributes, localPath);
                 objects_num++;
             }
             ClearProgressBar();
@@ -326,14 +328,14 @@ namespace ArenaUnity
             return string.IsNullOrWhiteSpace((string)el);
         }
 
-        private IEnumerator DownloadAssets(dynamic obj)
+        private IEnumerator DownloadAssets(string messageType, dynamic data)
         {
             string objUrl = null;
             string localPath = null;
             // update urls, if any
-            if (obj.type == "object")
+            if (messageType == "object")
             {
-                objUrl = ((string)obj.attributes.url).TrimStart('/');
+                objUrl = ((string)data.url).TrimStart('/');
                 if (objUrl.StartsWith("store/")) objUrl = $"https://{brokerAddress}/{objUrl}";
                 else if (objUrl.StartsWith("models/")) objUrl = $"https://{brokerAddress}/store/{objUrl}";
                 else objUrl = objUrl.Replace("www.dropbox.com", "dl.dropboxusercontent.com"); // replace dropbox links to direct links
@@ -391,20 +393,6 @@ namespace ArenaUnity
             yield return localPath;
         }
 
-        private void DisplayCancelableProgressBar(string title, string info, float progress)
-        {
-#if UNITY_EDITOR
-            EditorUtility.DisplayCancelableProgressBar(title, info, progress);
-#endif
-        }
-
-        private void ClearProgressBar()
-        {
-#if UNITY_EDITOR
-            EditorUtility.ClearProgressBar();
-#endif
-        }
-
         private static void SaveAsset(byte[] data, string path)
         {
             Directory.CreateDirectory(Path.GetDirectoryName(path));
@@ -420,9 +408,9 @@ namespace ArenaUnity
             }
         }
 
+        // methods for the editor
 #if UNITY_EDITOR
         [MenuItem("ARENA/Signout")]
-#endif
         internal static void SceneSignout()
         {
             EditorApplication.ExitPlaymode();
@@ -434,18 +422,15 @@ namespace ArenaUnity
         // Add a menu item to create custom GameObjects.
         // Priority 1 ensures it is grouped with the other menu items of the same kind
         // and propagated to the hierarchy dropdown and hierarchy context menus.
-#if UNITY_EDITOR
         [MenuItem("GameObject/ARENA/GLTF Model", false, 10)]
-#endif
         internal static void CreateArenaGltfModel(MenuCommand menuCommand)
         {
             ArenaObjectAddUrlWindow window = (ArenaObjectAddUrlWindow)EditorWindow.GetWindow(typeof(ArenaObjectAddUrlWindow));
             window.Init("gltf-model", menuCommand);
             window.Show();
         }
-#if UNITY_EDITOR
+
         [MenuItem("GameObject/ARENA/Image", false, 10)]
-#endif
         internal static void CreateArenaImage(MenuCommand menuCommand)
         {
             ArenaObjectAddUrlWindow window = (ArenaObjectAddUrlWindow)EditorWindow.GetWindow(typeof(ArenaObjectAddUrlWindow));
@@ -453,7 +438,18 @@ namespace ArenaUnity
             window.Show();
         }
 
-        private void CreateUpdateObject(string object_id, string storeType, dynamic data, string assetPath = null)
+        private void DisplayCancelableProgressBar(string title, string info, float progress)
+        {
+            EditorUtility.DisplayCancelableProgressBar(title, info, progress);
+        }
+
+        private void ClearProgressBar()
+        {
+            EditorUtility.ClearProgressBar();
+        }
+#endif
+
+        private void CreateUpdateObject(string object_id, string storeType, dynamic data, string assetPath = null, MenuCommand menuCommand = null)
         {
             ArenaObject aobj = null;
             if (arenaObjs.TryGetValue(object_id, out GameObject gobj))
@@ -473,6 +469,10 @@ namespace ArenaUnity
                             if (mobj != null)
                             {
                                 mobj.transform.parent = gobj.transform;
+                                foreach (Transform child in mobj.transform.GetComponentsInChildren<Transform>())
+                                {   // prevent inadvertent editing of gltf elements
+                                    child.gameObject.isStatic = true;
+                                }
                             }
                             break;
                         case "image":
@@ -495,6 +495,7 @@ namespace ArenaUnity
                 aobj.messageType = storeType;
                 aobj.parentId = (string)data.parent;
                 aobj.persist = true;
+                // camera
                 if ((string)data.object_type == "camera")
                 {   // sync camera to main display if requested
                     Camera cam = gobj.GetComponent<Camera>();
@@ -506,10 +507,19 @@ namespace ArenaUnity
                     else
                         cam.targetDisplay = ArenaUnity.secondDisplay;
                 }
+                // local create context auto-select
+                if (menuCommand != null)
+                {
+                    // Register the creation in the undo system
+                    Undo.RegisterCreatedObjectUndo(gobj, "Create " + gobj.name);
+                    Selection.activeObject = gobj;
+                }
             }
             // modify Unity attributes
             if (isElement(data.position))
                 gobj.transform.localPosition = ArenaUnity.ToUnityPosition(data.position);
+            else
+                gobj.transform.localPosition = Vector3.zero;
             if (isElement(data.rotation))
             {
                 // TODO: needed? bool invertY = !((string)data.object_type == "camera");
@@ -518,11 +528,15 @@ namespace ArenaUnity
                     gobj.transform.localRotation = ArenaUnity.ToUnityRotationQuat(data.rotation, invertY);
                 else // euler
                     gobj.transform.localRotation = ArenaUnity.ToUnityRotationEuler(data.rotation, invertY);
-                if ((string)data.object_type == "gltf-model")
-                    gobj.transform.rotation = ArenaUnity.GltfToUnityRotationQuat(gobj.transform.localRotation);
             }
+            else
+                gobj.transform.localRotation = Quaternion.identity;
+            if ((string)data.object_type == "gltf-model")
+                gobj.transform.localRotation = ArenaUnity.GltfToUnityRotationQuat(gobj.transform.localRotation);
             if (isElement(data.scale))
                 gobj.transform.localScale = ArenaUnity.ToUnityScale(data.scale);
+            else
+                gobj.transform.localScale = Vector3.one;
             if (isElement(data.material) || isElement(data.color))
                 ArenaUnity.ToUnityMaterial(data, ref gobj);
             ArenaUnity.ToUnityDimensions(data, ref gobj);
@@ -646,12 +660,12 @@ namespace ArenaUnity
             return csrfCookie;
         }
 
-        internal void Publish(string object_id, string msg)
+        internal void Publish(string object_id, string msgJson)
         {
-            byte[] payload = System.Text.Encoding.UTF8.GetBytes(msg);
+            byte[] payload = System.Text.Encoding.UTF8.GetBytes(msgJson);
             client.Publish($"{sceneTopic}/{client.ClientId}/{object_id}", payload, MqttMsgBase.QOS_LEVEL_AT_MOST_ONCE, false);
-            dynamic obj = JsonConvert.DeserializeObject(msg);
-            LogMessage("Sent", obj);
+            dynamic msg = JsonConvert.DeserializeObject(msgJson);
+            LogMessage("Sent", msg);
         }
 
         protected override void OnConnecting()
@@ -706,43 +720,42 @@ namespace ArenaUnity
             eventMessages.Add(eventMsg);
         }
 
-        internal void ProcessMessage(string msg)
+        internal void ProcessMessage(string msgJson, MenuCommand menuCommand = null)
         {
-            dynamic obj = JsonConvert.DeserializeObject(msg);
-            LogMessage("Received", obj);
-            StartCoroutine(ProcessArenaMessage(obj));
+            dynamic msg = JsonConvert.DeserializeObject(msgJson);
+            LogMessage("Received", msg);
+            StartCoroutine(ProcessArenaMessage(msg, menuCommand));
         }
 
-        private IEnumerator ProcessArenaMessage(dynamic obj)
+        private IEnumerator ProcessArenaMessage(dynamic msg, MenuCommand menuCommand = null)
         {
             // consume object updates
-            if (obj.type == "object")
+            if (msg.type == "object")
             {
-                switch ((string)obj.action)
+                switch ((string)msg.action)
                 {
                     case "create":
                     case "update":
-                        if (Convert.ToBoolean(obj.persist))
+                        if (Convert.ToBoolean(msg.persist))
                         {
                             string localPath = null;
-                            // TODO: fix coroutine error
-                            //if (isElement(obj.data) && isElement(obj.data.url) && !isElementEmpty(obj.data.url))
-                            //{
-                            //    DisplayCancelableProgressBar("ARENA Message", $"Loading object-id: {(string)obj.object_id}", 0f);
-                            //    CoroutineWithData cd = new CoroutineWithData(this, DownloadAssets(obj));
-                            //    yield return cd.coroutine;
-                            //    localPath = cd.result.ToString();
-                            //    ClearProgressBar();
-                            //}
-                            CreateUpdateObject((string)obj.object_id, (string)obj.type, obj.data, localPath);
+                            if (isElement(msg.data) && isElement(msg.data.url) && !isElementEmpty(msg.data.url))
+                            {
+                                DisplayCancelableProgressBar("ARENA Message", $"Loading object-id: {(string)msg.object_id}", 0f);
+                                CoroutineWithData cd = new CoroutineWithData(this, DownloadAssets((string)msg.type, msg.data));
+                                yield return cd.coroutine;
+                                localPath = cd.result.ToString();
+                                ClearProgressBar();
+                            }
+                            CreateUpdateObject((string)msg.object_id, (string)msg.type, msg.data, localPath, menuCommand);
                         }
-                        else if (obj.data.object_type == "camera") // try to manage camera
+                        else if (msg.data.object_type == "camera") // try to manage camera
                         {
-                            CreateUpdateObject((string)obj.object_id, (string)obj.type, obj.data);
+                            CreateUpdateObject((string)msg.object_id, (string)msg.type, msg.data);
                         }
                         break;
                     case "delete":
-                        RemoveObject((string)obj.object_id);
+                        RemoveObject((string)msg.object_id);
                         break;
                     default:
                         break;
@@ -751,17 +764,17 @@ namespace ArenaUnity
             }
         }
 
-        private void LogMessage(string dir, dynamic obj)
+        private void LogMessage(string dir, dynamic msg)
         {
             // determine logging level
-            if (!Convert.ToBoolean(obj.persist) && !logMqttNonPersist) return;
-            if (obj.type == "object")
+            if (!Convert.ToBoolean(msg.persist) && !logMqttNonPersist) return;
+            if (msg.type == "object")
             {
-                if (obj.data != null && obj.data.object_type == "camera" && !logMqttUsers) return;
+                if (msg.data != null && msg.data.object_type == "camera" && !logMqttUsers) return;
                 if (!logMqttObjects) return;
             }
-            if (obj.action == "clientEvent" && !logMqttEvents) return;
-            Debug.Log($"{dir}: {JsonConvert.SerializeObject(obj)}");
+            if (msg.action == "clientEvent" && !logMqttEvents) return;
+            Debug.Log($"{dir}: {JsonConvert.SerializeObject(msg)}");
         }
 
         protected void OnDestroy()