+++ Runtime/ArenaObjectAddUrlWindow.cs
@@ -1,4 +1,6 @@
+using System;
 using System.Dynamic;
+using System.Text.RegularExpressions;
 using Newtonsoft.Json;
 using UnityEditor;
 using UnityEngine;
@@ -25,35 +27,35 @@ namespace ArenaUnity
 
             if (GUILayout.Button($"Create new {object_type}"))
             {
+                if (ArenaClient.Instance == null)
+                {
+                    Debug.LogError($"Failed to create object '{object_id}', press Play before creating an ARENA {object_type}.");
+                    return;
+                }
+                // validate uri
+                if (!Uri.IsWellFormedUriString(object_url, UriKind.RelativeOrAbsolute))
+                {
+                    Debug.LogError($"Badly-formed Uri: '{object_url}'.");
+                    return;
+                }
                 dynamic msg = new ExpandoObject();
-                msg.object_id = object_id;
+                msg.object_id = Regex.Replace(object_id, ArenaUnity.regexArenaObjectId, "-"); ;
                 msg.action = "create";
                 msg.type = "object";
                 msg.persist = true;
                 dynamic data = new ExpandoObject();
                 data.object_type = object_type;
                 data.url = object_url;
+                Quaternion rotOut = object_type == "gltf-model" ? ArenaUnity.UnityToGltfRotationQuat(Quaternion.identity) : Quaternion.identity;
+                data.rotation = ArenaUnity.ToArenaRotationEuler(rotOut.eulerAngles);
                 msg.data = data;
                 string payload = JsonConvert.SerializeObject(msg);
-
-                if (ArenaClient.Instance != null)
-                {
-                    ArenaClient.Instance.Publish(object_id, payload); // remote
-                    ArenaClient.Instance.ProcessMessage(payload); // local
-
-                    GameObject go = GameObject.Find(object_id);
-                    // Ensure it gets reparented if this was a context click (otherwise does nothing)
-                    GameObjectUtility.SetParentAndAlign(go, menuCommand.context as GameObject);
-                    // Register the creation in the undo system
-                    Undo.RegisterCreatedObjectUndo(go, "Create " + go.name);
-                    Selection.activeObject = go;
-                }
-                else
-                    Debug.LogError($"Failed to create object '{object_id}', press Play before creating an ARENA {object_type}.");
+                ArenaClient.Instance.Publish(object_id, payload); // remote
+                ArenaClient.Instance.ProcessMessage(payload, menuCommand); // local
                 Close();
             }
 
-            if (GUILayout.Button("Abort"))
+            if (GUILayout.Button("Cancel"))
                 Close();
         }
     }