+++ Runtime/PrettyHierarchy/Editor/Utils/EditorUtils.cs
@@ -1,0 +1,11 @@
+﻿// From: https://github.com/NCEEGEE/PrettyHierarchy
+
+using UnityEditor;
+
+namespace PrettyHierarchy
+{
+    public static class EditorUtils
+    {
+        public static bool IsHierarchyFocused { get { return EditorWindow.focusedWindow != null && EditorWindow.focusedWindow.titleContent.text == "Hierarchy"; } }
+    }
+}